--- assets/stored_data.go
@@ -1,11 +1,0 @@
-package assets
-
-import . "broengine/datatypes"
-
-// Buggy when we uncomment these
-// var Sphere80 = ParseStl("sphere.stl", 1, 1, 1, 1)
-// var Cube12 = ParseStl("cube.stl", 1, 1, 1, 1)
-// var CubeRotated12 = ParseStl("cube_rotated.stl", 1, 1, 1, 1)
-
-var L1 = Light{Vector{0, 0, -50}, 30, 250, 3}
-var L2 = Light{Vector{0, 0, -50}, 15, 100, 3}
